Add umbrella protocols and implement ReadaheadTable methods
1. Add TreeNode which is an umbrella term form Tree and Token
When tables run, they add tree AND token to the treeStack.  To make treeStack store both types, I added a TreeNode protocol that refers to both Token and Tree.

2. Make TableWithTransitions a protocol which has two other types
Transitions can either have a key of integer and/or string but Swift does not allow us to mix both in one Dictionary. I decided to make two different types of tables to avoid making a new class that is Hashable (requirement for being a Dictionary key) . If needed, we can change the implementation in the future.